Tcrud v0.2 Response / NestedReply

+++ src/main/java/com/example/tcrud/dto/BoardDto.java
@@ -41,7 +41,7 @@ public class BoardDto {
         private Long id;
         private String title;
         private String content;
-        private Integer replycnt;
+        private Long replycnt;
         private Long viewcnt;
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
         private LocalDateTime insertTime;
